feat: remove stopwords removal
I'll be using BERT based embedding models and removing stopwords is not
always necessary (it may sometimes hurt to do so).
At first the stopwords removal was introduced as a way to improve
embedding clustering via agglomerative clustering.
I'll be changing this approach to umap (for dimensionality reduction)
plus pynndescent for fast KNN classification.